backend/requirements.txt
#
#  requirements.txt
#  TapInApp - Wordle Leaderboard Server
#
#  Created by Darius Ehsani on 2/2/26.
#
#  MARK: - Python Dependencies
#  Install with: pip install -r requirements.txt
#

# Flask web framework
Flask==3.0.0

# CORS support for cross-origin requests from iOS app
Flask-Cors==4.0.0

# WSGI server for production (used by Google App Engine)
gunicorn==21.2.0

# Anthropic Claude API client
anthropic>=0.39.0

# Environment variable loading from .env files
python-dotenv>=1.0.0

# Google Cloud Firestore (persistent storage for processed events)
google-cloud-firestore>=2.14.0

# RSS feed parsing (for The Aggie article caching)
feedparser>=6.0.0

# HTML parsing for article body extraction (server-side)
beautifulsoup4>=4.12.0

# JWT token creation and verification (user auth)
PyJWT[crypto]>=2.8.0

# Password hashing (email/password auth)
bcrypt>=4.1.0

# HTTP requests (Apple JWKS fetch)
requests>=2.31.0

# Firebase Admin SDK (phone auth token verification)
firebase-admin>=6.0.0
